Add more realistic test data
